+++ MAPL_Base/MAPL_Cap.F90
@@ -14,11 +14,10 @@ module MAPL_CapMod
    use MAPL_BaseMod
    use MAPL_ExceptionHandling
    use pFIO
-   use MAPL_Profiler, only: get_global_time_profiler, BaseProfiler, TimeProfiler
    use MAPL_ioClientsMod
    use MAPL_CapOptionsMod
-   use pflogger, only: logging
-   use pflogger, only: Logger
+   use MAPL_ServerManager
+   use MAPL_ApplicationSupport
    implicit none
    private
 
@@ -37,11 +36,8 @@ module MAPL_CapMod
       logical :: mpi_already_initialized = .false.
 
       type(MAPL_CapGridComp), public :: cap_gc
-      type(SplitCommunicator)  :: split_comm
       type(MAPL_Communicators) :: mapl_comm
-      type(MpiServer), pointer :: i_server=>null()
-      type(MpiServer), pointer :: o_server=>null()
-      type(DirectoryService) :: directory_service
+      type(ServerManager) :: cap_server
 
    contains
       procedure :: run
@@ -90,7 +86,6 @@ contains
       class ( MAPL_CapOptions), optional, intent(in) :: cap_options
       integer, optional, intent(out) :: rc
       integer :: status
-      type(Logger), pointer :: lgr
 
       cap%name = name
       cap%set_services => set_services
@@ -111,59 +106,14 @@ contains
       call cap%initialize_mpi(rc=status)
       _VERIFY(status)
 
-      call initialize_pflogger()
+      call MAPL_Initialize(comm=cap%comm_world, &
+                           logging_config=cap%cap_options%logging_config, &
+                           rc=status)
+      _VERIFY(status)
 
       _RETURN(_SUCCESS)     
       _UNUSED_DUMMY(unusable)
 
-    contains
-
-       subroutine initialize_pflogger()
-          use pflogger, only: pfl_initialize => initialize
-          use pflogger, only: StreamHandler, FileHandler, HandlerVector
-          use pflogger, only: MpiLock, MpiFormatter
-          use pflogger, only: INFO, WARNING
-          use, intrinsic :: iso_fortran_env, only: OUTPUT_UNIT
-          type (HandlerVector) :: handlers
-          type (StreamHandler) :: console
-          type (FileHandler) :: file_handler
-          integer :: level
-          
-          call pfl_initialize()
-
-          if (cap%cap_options%logging_config /= '') then
-             call logging%load_file(cap%cap_options%logging_config)
-          else
-
-             console = StreamHandler(OUTPUT_UNIT)
-             call console%set_level(INFO)
-             call console%set_formatter(MpiFormatter(cap%comm_world, fmt='%(short_name)a10~: %(message)a'))
-             call handlers%push_back(console)
-
-             file_handler = FileHandler('warnings_and_errors.log')
-             call file_handler%set_level(WARNING)
-             call file_handler%set_formatter(MpiFormatter(cap%comm_world, fmt='pe=%(mpi_rank)i5.5~: %(short_name)a~: %(message)a'))
-             call file_handler%set_lock(MpiLock(cap%comm_world))
-             call handlers%push_back(file_handler)
-             
-             if (cap%rank == 0) then
-                level = INFO
-             else
-                level = WARNING
-             end if
-
-             call logging%basic_config(level=level, handlers=handlers, rc=status)
-             _VERIFY(status)
-             
-             if (cap%rank == 0) then
-                lgr => logging%get_logger('MAPL')
-                call lgr%warning('No configure file specified for logging layer.  Using defaults.')
-             end if
-
-          end if
-
-       end subroutine initialize_pflogger
-     
     end function new_MAPL_Cap
 
    
@@ -196,14 +146,17 @@ contains
 
       integer :: status
       integer :: subcommunicator
+      type(SplitCommunicator) :: split_comm
 
       _UNUSED_DUMMY(unusable)
       
       subcommunicator = this%create_member_subcommunicator(this%comm_world, rc=status); _VERIFY(status)
       if (subcommunicator /= MPI_COMM_NULL) then
          call this%initialize_io_clients_servers(subcommunicator, rc = status); _VERIFY(status)
+         call this%cap_server%get_splitcomm(split_comm)
+         call fill_mapl_comm(split_comm, subcommunicator, .false., this%mapl_comm, rc=status)
          call this%run_member(rc=status); _VERIFY(status)
-         call this%directory_service%free_directory_resources()
+         call this%cap_server%finalize()
       end if
 
       _RETURN(_SUCCESS)
@@ -212,144 +165,24 @@ contains
 
 
    subroutine initialize_io_clients_servers(this, comm, unusable, rc)
-     use MAPL_CFIOMod
      class (MAPL_Cap), target, intent(inout) :: this
      integer, intent(in) :: comm
      class (KeywordEnforcer), optional, intent(in) :: unusable
      integer, optional, intent(out) :: rc
-
-     type (SimpleCommSplitter) :: splitter
-     integer :: status, i, rank
-     logical :: running_old_o_server ! relevant only for "old" o-server
-     character(len=:), allocatable :: s_name
-
-     type(ClientThread), pointer :: clientPtr
+     integer :: status
 
      _UNUSED_DUMMY(unusable)
-
-     this%directory_service = DirectoryService(comm)
-     splitter = SimpleCommSplitter(comm)
-     call splitter%add_group(npes=this%cap_options%npes_model, name='model', isolate_nodes=.true.)
-
-     if (this%cap_options%npes_input_server(1) > 0) then
-        do i = 1, this%cap_options%n_iserver_group
-           s_name ='i_server'//trim(i_to_string(i))
-           call splitter%add_group(npes=this%cap_options%npes_input_server(i), name=s_name, isolate_nodes=.true.)
-        enddo
-     elseif (this%cap_options%nodes_input_server(1) > 0) then
-        do i = 1, this%cap_options%n_iserver_group
-           s_name ='i_server'//trim(i_to_string(i))
-           call splitter%add_group(nnodes=this%cap_options%nodes_input_server(i), name=s_name, isolate_nodes=.true.)
-        enddo
-     end if
-
-     running_old_o_server = .false.
-
-     if (this%cap_options%npes_output_server(1) > 0 ) then
-        running_old_o_server = (this%cap_options%n_oserver_group ==1) ! otherwise need to combine all the o-server groups to form a group used by old_server
-        do i = 1, this%cap_options%n_oserver_group
-          s_name ='o_server'//trim(i_to_string(i))
-          call splitter%add_group(npes=this%cap_options%npes_output_server(i), name=s_name, isolate_nodes=.true.)
-        enddo
-     else if(this%cap_options%nodes_output_server(1) > 0) then
-        running_old_o_server = (this%cap_options%n_oserver_group ==1) ! otherwise need to combine all the o-server groups to form a group used by old_server
-        do i = 1, this%cap_options%n_oserver_group
-          s_name ='o_server'//trim(i_to_string(i))
-          call splitter%add_group(nnodes=this%cap_options%nodes_output_server(i), name=s_name, isolate_nodes=.true.)
-        enddo
-     endif
-
-     this%split_comm = splitter%split(rc=status); _VERIFY(status)
-
-     call fill_mapl_comm(this%split_comm, comm, running_old_o_server, this%mapl_comm, rc=status)
+     call this%cap_server%initialize(comm, &
+         application_size=this%cap_options%npes_model, &
+         nodes_input_server=this%cap_options%nodes_input_server, &
+         nodes_output_server=this%cap_options%nodes_output_server, &
+         npes_input_server=this%cap_options%npes_input_server, &
+         npes_output_server=this%cap_options%npes_output_server, &
+         rc=status)
      _VERIFY(status)
-     
-     s_name = this%split_comm%get_name()
 
-     if ( index(s_name, 'model') /=0 ) then
-        if (this%cap_options%npes_input_server(1) == 0 .and. this%cap_options%nodes_input_server(1) == 0) then
-           allocate(this%i_server, source = MpiServer(this%split_comm%get_subcommunicator(), 'i_server'//trim(i_to_string(1))))
-           call this%directory_service%publish(PortInfo('i_server'//trim(i_to_string(1)), this%i_server), this%i_server)
-        end if
-        if (this%cap_options%npes_output_server(1) == 0 .and. this%cap_options%nodes_output_server(1) == 0) then
-           allocate(this%o_server, source = MpiServer(this%split_comm%get_subcommunicator(), 'o_server'//trim(i_to_string(1))))
-           call this%directory_service%publish(PortInfo('o_server'//trim(i_to_string(1)), this%o_server), this%o_server)
-        end if
-        call io_client%init_io_clients(ni = this%cap_options%n_iserver_group, no = this%cap_options%n_oserver_group )
-     endif
-
-     ! establish i_server group one by one
-     do i = 1, this%cap_options%n_iserver_group
-        
-        if ( trim(s_name) =='i_server'//trim(i_to_string(i)) ) then
-           allocate(this%i_server, source = MpiServer(this%split_comm%get_subcommunicator(), s_name))
-           call this%directory_service%publish(PortInfo(s_name,this%i_server), this%i_server)
-           call this%directory_service%connect_to_client(s_name, this%i_server)
-           call MPI_Comm_Rank(this%split_comm%get_subcommunicator(),rank,status)
-           if (rank == 0 .and. this%cap_options%nodes_input_server(i) /=0 ) then
-              write(*,'(A,I0,A)')"Starting pFIO input server on ",this%cap_options%nodes_input_server(i)," nodes"
-           else if (rank==0 .and. this%cap_options%npes_input_server(i) /=0 ) then
-              write(*,'(A,I0,A)')"Starting pFIO input server on ",this%cap_options%npes_input_server(i)," pes"
-           end if
-        endif
-
-        if ( index(s_name, 'model') /=0 ) then
-           clientPtr => i_Clients%current()
-           call this%directory_service%connect_to_server('i_server'//trim(i_to_string(i)), clientPtr, this%split_comm%get_subcommunicator())
-           call i_Clients%next()
-        endif 
-
-        call mpi_barrier(comm, status) 
-
-     enddo
-
-     ! establish o_server group one by one
-     do i = 1, this%cap_options%n_oserver_group
-        
-        if ( trim(s_name) =='o_server'//trim(i_to_string(i)) ) then
-           allocate(this%o_server, source = MpiServer(this%split_comm%get_subcommunicator(), s_name))
-           call this%directory_service%publish(PortInfo(s_name,this%o_server), this%o_server)
-           call this%directory_service%connect_to_client(s_name, this%o_server)
-           call MPI_Comm_Rank(this%split_comm%get_subcommunicator(),rank,status)
-           if (rank == 0 .and. this%cap_options%nodes_output_server(i) /=0 ) then
-              write(*,'(A,I0,A)')"Starting pFIO output server on ",this%cap_options%nodes_output_server(i)," nodes"
-           else if (rank==0 .and. this%cap_options%npes_output_server(i) /=0 ) then
-              write(*,'(A,I0,A)')"Starting pFIO output server on ",this%cap_options%npes_output_server(i)," pes"
-           end if
-        endif
-
-        if ( index(s_name, 'model') /=0 ) then
-           clientPtr => o_Clients%current()
-           call this%directory_service%connect_to_server('o_server'//trim(i_to_string(i)), clientPtr, this%split_comm%get_subcommunicator())
-           call o_Clients%next()
-        endif 
-
-        call mpi_barrier(comm, status) 
-
-     enddo
-
-     if ( index(s_name, 'o_server') /=0 ) then
-        call this%o_server%start()
-     endif
-
-     if ( index(s_name, 'i_server') /=0 ) then
-        call this%i_server%start()
-     endif
-
-     if ( index(s_name, 'model') /=0 ) then
-        call i_Clients%set_current(1) ! set current to be the first
-        call o_Clients%set_current(1) ! set current to be the first
-        if (this%cap_options%npes_output_server(1) >0) then
-           call io_client%set_size(no = this%cap_options%npes_output_server,rc=status)
-        else if (this%cap_options%nodes_output_server(1)>0) then
-           call io_client%set_size(no = this%cap_options%nodes_output_server,rc=status)
-        endif
-        _VERIFY(status)
-     end if
-     
    end subroutine initialize_io_clients_servers
-
-   
+     
    ! This layer splits the communicator to support separate i/o servers
    ! and runs the model via a CapGridComp.
    subroutine run_member(this, rc)
@@ -358,8 +191,10 @@ contains
       integer, optional, intent(out) :: rc
       
       integer :: status
+      type(SplitCommunicator) :: split_comm
 
-      select case(this%split_comm%get_name())
+      call this%cap_server%get_splitcomm(split_comm)
+      select case(split_comm%get_name())
       case('model')
          call this%run_model(this%mapl_comm, rc=status); _VERIFY(status)
          call i_Clients%terminate()
@@ -473,16 +308,8 @@ contains
       type (ESMF_VM) :: vm
       integer :: start_tick, stop_tick, tick_rate
       integer :: status
-!
-!     profiler
-!  
-      class (BaseProfiler), pointer :: t_p
 
       _UNUSED_DUMMY(unusable)
-
-      t_p => get_global_time_profiler()
-      t_p = TimeProfiler('All', comm_world = mapl_comm%esmf%comm)
-      call t_p%start()
 
       call start_timer()
       call ESMF_Initialize (vm=vm, logKindFlag=this%cap_options%esmf_logging_mode, mpiCommunicator=mapl_comm%esmf%comm, rc=status)
@@ -503,8 +330,6 @@ contains
       !_VERIFY(status)
       call stop_timer()
 
-      call t_p%stop()
-      call report_profiling()
       ! W.J note : below reporting will be remove soon
       call report_throughput()
 
@@ -539,56 +364,6 @@ contains
          end if
          
       end subroutine report_throughput
-
-      subroutine report_profiling(rc)
-         use MAPL_Profiler
-         integer, optional, intent(out) :: rc
-         type (ProfileReporter) :: reporter
-         integer :: i
-         character(:), allocatable :: report_lines(:)
-         type (MultiColumn) :: inclusive
-         type (MultiColumn) :: exclusive
-         integer :: npes, my_rank, rank, ierror
-         character(1) :: empty(0)
-
-         reporter = ProfileReporter(empty)
-         call reporter%add_column(NameColumn(50, separator= " "))
-         call reporter%add_column(FormattedTextColumn('#-cycles','(i5.0)', 5, NumCyclesColumn(),separator='-'))
-
-         inclusive = MultiColumn(['Inclusive'], separator='=')
-         call inclusive%add_column(FormattedTextColumn(' T (sec) ','(f9.3)', 9, InclusiveColumn(), separator='-'))
-         call inclusive%add_column(FormattedTextColumn('   %  ','(f6.2)', 6, PercentageColumn(InclusiveColumn(),'MAX'),separator='-'))
-         call reporter%add_column(inclusive)
-
-         exclusive = MultiColumn(['Exclusive'], separator='=')
-         call exclusive%add_column(FormattedTextColumn(' T (sec) ','(f9.3)', 9, ExclusiveColumn(), separator='-'))
-         call exclusive%add_column(FormattedTextColumn('   %  ','(f6.2)', 6, PercentageColumn(ExclusiveColumn()), separator='-'))
-         call reporter%add_column(exclusive)
-
-!!$      call reporter%add_column(FormattedTextColumn('  std. dev ','(f12.4)', 12, StdDevColumn()))
-!!$      call reporter%add_column(FormattedTextColumn('  rel. dev ','(f12.4)', 12, StdDevColumn(relative=.true.)))
-!!$      call reporter%add_column(FormattedTextColumn('  max cyc ','(f12.8)', 12, MaxCycleColumn()))
-!!$      call reporter%add_column(FormattedTextColumn('  min cyc ','(f12.8)', 12, MinCycleColumn()))
-!!$      call reporter%add_column(FormattedTextColumn(' mean cyc','(f12.8)', 12, MeanCycleColumn()))
-!!$         call mem_reporter%add_column(NameColumn(50,separator='-'))
-!!$         call mem_reporter%add_column(MemoryTextColumn(['RSS'],'(i10,1x,a2)',13, InclusiveColumn(),separator='-'))
-!!$         call mem_reporter%add_column(MemoryTextColumn(['Cyc RSS'],'(i10,1x,a2)',13, MeanCycleColumn(),separator='-'))
-
-!!$         report_lines = reporter%generate_report(get_global_time_profiler())
-
-         call MPI_Comm_size(mapl_comm%esmf%comm, npes, ierror)
-         call MPI_Comm_Rank(mapl_comm%esmf%comm, my_rank, ierror)
-
-         if (my_rank == 0) then
-               report_lines = reporter%generate_report(t_p)
-               write(*,'(a,1x,i0)')'Report on process: ', my_rank
-               do i = 1, size(report_lines)
-                  write(*,'(a)') report_lines(i)
-               end do
-          end if
-          call MPI_Barrier(mapl_comm%esmf%comm, ierror)
-
-      end subroutine report_profiling
 
    end subroutine run_model
    
@@ -699,7 +474,7 @@ contains
       _UNUSED_DUMMY(unusable)
 
       if (.not. this%mpi_already_initialized) then
-         call logging%free()
+         call MAPL_Finalize(comm=this%comm_world)
          call MPI_Finalize(ierror)
          _VERIFY(ierror)
       end if